requirements.txt
numpy
python-mnist
progress